清除工作数据库中多余的证券名称项，更新测试数据库。 初步完成。

+++ src/Data Container/ContainerFinnhubStock.cpp
@@ -245,7 +245,7 @@ void CContainerFinnhubStock::UpdateBasicFinancialMetricDB(const vector<CFinnhubS
 	// 添加新的基本财务数据
 	if (iCurrentUpdated < iBasicFinancialNeedUpdate) {
 		ASSERT(setBasicFinancialMetric.IsEOF());
-		for (int i = 0; i < iBasicFinancialNeedUpdate; i++) {
+		for (size_t i = 0; i < iBasicFinancialNeedUpdate; i++) {
 			const auto& pStockNeedAppend = vStock.at(i);
 			if (pStockNeedAppend->IsUpdateBasicFinancialDB()) {
 				iCurrentUpdated++;
